refactor(api): move configuration interaces to domain
