+++ Src/Campus.Domain.Core/Interfaces/IConfigurationProviderFactory.cs
@@ -1,4 +1,4 @@
-namespace Campus.Master.API.Helpers.Contracts
+namespace Campus.Domain.Core.Interfaces
 {
     public interface IConfigurationProviderFactory
     {
