+++ Src/Campus.Master.API/Helpers/Implementations/ConfigurationProviderFactory.cs
@@ -1,6 +1,6 @@
-using Campus.Master.API.Helpers.Contracts;
+using Campus.Domain.Core.Interfaces;
 using Microsoft.Extensions.Configuration;
-using IConfigurationProvider = Campus.Master.API.Helpers.Contracts.IConfigurationProvider;
+using ICampusConfigurationProvider = Campus.Domain.Core.Interfaces.IConfigurationProvider;
 
 namespace Campus.Master.API.Helpers.Implementations
 {
@@ -15,7 +15,7 @@ namespace Campus.Master.API.Helpers.Implementations
             InDevelopmentMode = inDevelopmentMode;
         }
         
-        public IConfigurationProvider CreateProvider()
+        public ICampusConfigurationProvider CreateProvider()
         {
             if (InDevelopmentMode)
                 return new AppSettingsConfigurationProvider(Configuration);
